multimodal_insights_app/backend/requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-multipart==0.0.12
aiofiles==24.1.0
websockets==13.1

# Logging & Telemetry
structlog==24.4.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
azure-monitor-opentelemetry-exporter==1.0.0b31

# Microsoft Agent Framework
agent-framework  # Microsoft Agent Framework for multi-agent orchestration

# Azure AI Services
openai==1.51.0
azure-identity==1.19.0
azure-ai-documentintelligence==1.0.0b4  # Document Intelligence SDK v4.0
azure-cognitiveservices-speech==1.40.0  # Speech Services
azure-storage-blob==12.23.1
azure-cosmos==4.7.0

# Data Processing
pandas==2.2.3
numpy==2.1.2

# File Processing
PyPDF2==3.0.1
pypdf==5.0.1
python-magic-bin==0.4.14  # File type detection (Windows)
pillow==10.4.0  # Image processing
moviepy==1.0.3  # Video processing
pydub==0.25.1  # Audio format conversion

# PDF Export
reportlab==4.2.5
markdown==3.7
pdfkit==1.0.0  # HTML to PDF (requires wkhtmltopdf installed)

# HTTP & Requests
requests==2.32.3
httpx==0.27.2

# Utilities
tabulate==0.9.0
python-dateutil==2.9.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
black==24.10.0
ruff==0.7.3
